TestPersistentMappedEBO を修正

+++ render/src/main/java/me/hannsi/lfjg/render/system/mesh/persistent/TestPersistentMappedEBO.java
@@ -114,7 +114,7 @@ public class TestPersistentMappedEBO implements TestPersistentMappedBuffer {
             newCapacity = requiredBytes;
         }
 
-        growBuffer((int) newCapacity);
+        growBuffer(newCapacity);
     }
 
     @Override
